moved the notion code to notion.js
