+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@notionhq/client": "^2.2.14",
     "dot-env": "^0.0.1"
   }
 }
